extract worldstateinterpolator
