+++ src/Game/Components/Components.js
@@ -1,7 +1,6 @@
 /** @format */
 
 export { WebSocketService } from './WebSocketService';
-export { WorldStateInterpolator } from './WorldStateInterpolator';
+export { WorldStateInterpolator } from './StateComputer/WorldStateInterpolator';
 export { AssetsManager } from './AssetsManager';
 export { InputManager } from './InputManager';
-
